+++ java/tf2/items/guns/ItemTFGunsSG.java
@@ -219,7 +219,7 @@ public class ItemTFGunsSG extends ItemTFGuns implements IHasModel
 		double i = this.damage(stack);
 		int x = this.getMaxDamage(stack);
 		double y = this.reloadTime(stack) / 20;
-		double w = this.recoil + this.recoilrand;
+		float w = this.recoil + this.recoilrand;
 		int z = this.coolTime(stack);
 
 		tooltip.add(ChatFormatting.BLUE + I18n.translateToLocal(" ") + (i) + "x" + (k) + " " + I18n.translateToLocal("tf.bulletdamage"));
